chore: clippy cleanup — dead code, unnecessary casts, unused mut
- Remove dead SHELL_EXPANSION_PATTERNS constant
- Refactor 3 risk_level assignments (match-as-expression pattern)
- Mark public_bytes() as #[cfg(test)] in crypto
- Fix 7 unnecessary type casts in billing (u64/u32)
- Use div_ceil() in invoice calculation
- Restore Threat/ThreatLevel re-exports (broken by clippy --fix)
- 28 files from previous cargo fix (unused imports)

All tests passing, 0 failures.

+++ crates/billing/src/lib.rs
@@ -25,7 +25,6 @@ use anyhow::Result;
 use chrono::{DateTime, Utc};
 use dashmap::DashMap;
 use serde::{Deserialize, Serialize};
-use tracing;
 
 /// Billing engine — central struct for all billing operations
 pub struct BillingEngine {
@@ -258,10 +257,10 @@ impl BillingEngine {
         };
 
         let current_value = match operation {
-            usage::UsageOperation::ApiRequest => current.api_requests_today as u64 + 1,
-            usage::UsageOperation::Tokens(n) => current.tokens_today as u64 + n as u64,
-            usage::UsageOperation::AgentConnect => current.active_agents as u64 + 1,
-            usage::UsageOperation::StorageBytes(n) => current.storage_bytes as u64 + n as u64,
+            usage::UsageOperation::ApiRequest => current.api_requests_today + 1,
+            usage::UsageOperation::Tokens(n) => current.tokens_today + n,
+            usage::UsageOperation::AgentConnect => current.active_agents + 1,
+            usage::UsageOperation::StorageBytes(n) => current.storage_bytes + n,
         };
 
         if limit > 0 && current_value > limit {
